made minor changes to include adaption of end points in the .cpp files and added functionality to plot rel error

+++ project1/code/delete_files.py
@@ -1,5 +1,6 @@
 from os import system
 import sys
+import os.path
 
 try:
     algo = sys.argv[1]
@@ -9,5 +10,8 @@ except:
 
 # Deleting files
 for i in range(1, n+1):
-    system("rm ./output/" + algo + str(i))
     system("rm ./output/timed_" + algo + str(i))
+    if not os.path.isfile("./output/" + algo + str(i)):
+        continue
+    else:
+        system("rm ./output/" + algo + str(i))
